feat(null-filter): initial implementation
It is known to not work in all cases, i.e. it can only take one
key-value pair at a time (no consecutive ones), but besides that
is a pretty optimal implementation (even though it aint a pretty one).

However, the test still fails, we match nothing for some reason.
Must be evaluated later.